+++ bb.ps1
@@ -1,3 +1,0 @@
-Set-Alias run ./bb.ps1
-./build.bat
-./molvis.exe