+++ run_windows.ps1
@@ -1,0 +1,3 @@
+Set-Alias run ./run_windows.ps1
+./build.bat
+./molvis.exe